Split out daily edition metrics spec
We want to start testing that editions are part of
the data. We are going to start with the API
endpoints.

We are splitting out into a separate spec to make
our existing tests simpler, and also for the
intention to be clearer.
